Fit an additive linear mixed effect model over replicates

ngme function performs an analysis of non-gaussian additive models. It does the maximum likelihood estimation via stochastic gradient descent. The prediction of unknown location can be performed by leaving the response variable to be NA. The likelihood is specified by family. The model estimation control can be setted in control using control_opt() function, see ?control_opt for details. See ngme_model_types() for available models.

Usage

ngme(
  formula,
  data,
  family = "normal",
  control_opt = NULL,
  control_ngme = NULL,
  group = NULL,
  replicate = NULL,
  start = NULL,
  moving_window = 1,
  debug = FALSE
)

Arguments

formula

formula

data

a dataframe or a list providing data (Only response variable can contain NA value, NA value in other columns will cause problem)

family

likelihood type, same as measurement noise specification, 1. string 2. ngme noise obejct

control_opt

control for optimizer. by default it is control_opt(). See ?control_opt for details.

control_ngme

control for ngme model. by default it is control_ngme(). See ?control_ngme for details.

group

factor, used for bivariate model, indicating which group the observation belongs to

replicate

factor, used for divide data into different replicates

start

starting ngme object (usually object from last fit)

moving_window

number of iterations to average the estimation

debug

toggle debug mode

Value

random effects (for different replicate) + models(fixed effects, measuremnt noise, and latent process)

Examples

ngme(
  formula = Y ~ x1 + f(
    x2,
    model = "ar1",
    noise = noise_nig(),
    rho = 0.5
  ) + f(x1,
    model = "rw1",
    noise = noise_normal(),
  ),
  family = noise_normal(sd = 0.5),
  data = data.frame(Y = 1:5, x1 = 2:6, x2 = 3:7),
  control_opt = control_opt(
    estimation = FALSE
  )
)
#> Using character strings for 'model' is not recommended (will be deprecated in the future). Please consider using operator functions like matern(), ar1(), etc. (e.g., model = matern()).
#> Error in (function (mesh = NULL, rho = 0) {    if (is.null(mesh)) {        return(structure(list(model = "ar1", args = list(rho = rho)),             class = "ngme_operator_def"))    }    stopifnot(`rho should be between -1 and 1` = rho >= -1 &&         rho <= 1)    mesh <- ngme_build_mesh(mesh)    n <- mesh$n    h <- c(diff(mesh$loc), 1)    G <- Matrix::Diagonal(n)    G[1, 1] <- sqrt(1 - rho^2)    C <- Matrix::sparseMatrix(j = 1:(n - 1), i = 2:n, x = -1,         dims = c(n, n))    G <- ngme_as_sparse(G)    C <- ngme_as_sparse(C)    stopifnot(`The mesh should be 1d and has gap 1.` = all(h ==         1))    theta_K <- ar1_a2th(rho)    stopifnot(`The length of rho(theta_K) should be 1.` = length(theta_K) ==         1)    update_K <- function(theta_K) {        ar1_th2a(theta_K) * C + G    }    g <- name2fun("tanh", inv = TRUE)    ngme_operator(mesh = mesh, model = "ar1", theta_K = c(rho = g(rho)),         trans = c(rho = "tanh"), matrices = list(C, G), h = h,         C = C, G = G, update_K = update_K, K = ngme_as_sparse(update_K(theta_K)),         symmetric = FALSE, zero_trace = FALSE, param_name = "rho",         param_trans = list(ar1_th2a), generic_type = "generic")})(model_name = "ar1", map = 3:7, model = "ar1", noise = structure(list(    noise_type = "nig", V = NULL, theta_mu = 0, theta_sigma = 0,     theta_sigma_normal = numeric(0), theta_nu = 0, B_mu = structure(1, dim = c(1L,     1L)), B_sigma = structure(1, dim = c(1L, 1L)), B_sigma_normal = structure(logical(0), dim = 1:0),     B_nu = structure(1, dim = c(1L, 1L)), n_theta_mu = 1L, n_theta_sigma = 1L,     n_theta_nu = 1L, n_rho = 0L, n_theta_sigma_normal = 0L, fix_theta_mu = FALSE,     fix_theta_sigma = FALSE, fix_theta_nu = FALSE, nu_lower_bound = 0.01,     fix_V = FALSE, fix_rho = FALSE, fix_theta_sigma_normal = FALSE,     n_params = 3L, single_V = FALSE, share_V = FALSE, corr_measurement = FALSE,     index_corr = NULL, map_corr = NULL, rho = numeric(0), prior_mu = structure(list(        type = "normal", param = c(0, 0.01)), class = "ngme_prior"),     prior_sigma = structure(list(type = "normal", param = c(0,     0.01)), class = "ngme_prior"), prior_nu = structure(list(        type = "normal", param = c(0, 0.01)), class = "ngme_prior")), class = "ngme_noise"),     mesh = structure(list(n = 5L, m = 5, loc = 3:7, mid = 3:7,         interval = c(3L, 7L), boundary = c("neumann", "neumann"        ), cyclic = FALSE, manifold = "R1", degree = 1, free.clamped = c(FALSE,         FALSE), idx = list(loc = 1:5)), class = c("fm_mesh_1d",     "inla.mesh.1d")), name = "field1", data = structure(list(        Y = 1:5, x1 = 2:6, x2 = 3:7), class = "data.frame", row.names = c(NA,     -5L)), group = structure(c(1L, 1L, 1L, 1L, 1L), levels = "1", class = "factor"),     fix_W = FALSE, fix_theta_K = FALSE, prior_theta_K = structure(list(        type = "normal", param = c(0, 0.001)), class = "ngme_prior"),     subset = c(TRUE, TRUE, TRUE, TRUE, TRUE), debug = FALSE,     rho = 0.5): unused arguments (model_name = "ar1", map = 3:7, model = "ar1", noise = list("nig", NULL, 0, 0, numeric(0), 0, 1, 1, logical(0), 1, 1, 1, 1, 0, 0, FALSE, FALSE, FALSE, 0.01, FALSE, FALSE, FALSE, 3, FALSE, FALSE, FALSE, NULL, NULL, numeric(0), list("normal", c(0, 0.01)), list("normal", c(0, 0.01)), list("normal", c(0, 0.01))), name = "field1", data = list(1:5, 2:6, 3:7), group = c(1, 1, 1, 1, 1), fix_W = FALSE, fix_theta_K = FALSE, prior_theta_K = list("normal", c(0, 0.001)), subset = c(TRUE, TRUE, TRUE, TRUE, 
#> TRUE), debug = FALSE)