Generate control specifications for ngme() function.

These are configurations for ngme estimation and optimization process.

Usage

ngme_control(
  burnin = 100,
  iterations = 100,
  gibbs_sample = 5,
  stepsize = 1,
  estimation = TRUE,
  n_parallel_chain = 2,
  stop_points = 10,
  exchange_VW = TRUE,
  n_slope_check = 3,
  std_lim = 0.1,
  trend_lim = 0.05,
  print_check_info = TRUE,
  opt_beta = TRUE,
  fix_beta = FALSE,
  max_relative_step = 0.1,
  max_absolute_step = 0.5,
  reduce_var = FALSE,
  reduce_power = 0.75,
  threshold = 1e-05,
  window_size = 1
)

Arguments

burnin

burn-in periods

iterations

optimizing terations

gibbs_sample

number of gibbs sampels

stepsize

stepsize

estimation

estimating the parameters

n_parallel_chain

number of parallel chains

stop_points

number of stop points for convergence check

exchange_VW

exchange last V and W in each chian

n_slope_check

number of stop points for regression

std_lim

maximum allowed standard deviation

trend_lim

maximum allowed slope

print_check_info

print the convergence information

opt_beta

logical, optimize fixed effect

fix_beta

logical, fix fixed effect

max_relative_step

max relative step allowed in 1 iteration

max_absolute_step

max absolute step allowed in 1 iteration

reduce_var

logical, reduce variace

reduce_power

numerical the power of reduce level

threshold

till when start to reduce the variance

window_size

numerical, length of window for final estimates

Value

list of control variables

Details

To enable convergence check, we need multiple chains running. We compare the trend of the estimated parameter of length n_slope_check (linear regression) with trend_lim. We compare the standard devation of estimated parameters (in different chains) with std_lim.