ngme space-time non-separable model specification

spacetime(
  mesh,
  lambda = 1,
  alpha = 2,
  fix_gamma = FALSE,
  theta_gamma_x = 0,
  theta_gamma_y = 0,
  B_gamma_x = matrix(1, nrow = mesh[[2]]$n, ncol = 1),
  B_gamma_y = matrix(1, nrow = mesh[[2]]$n, ncol = 1),
  B_gamma_x_list = NULL,
  B_gamma_y_list = NULL,
  cc = 1,
  kappa = 1,
  stabilization = TRUE,
  ...
)

Arguments

mesh

A list of two objects: - `mesh_t`: The temporal mesh. - `mesh_s`: The spatial mesh.

lambda

The spatial damping parameter.

alpha

2 or 4, SPDE smoothness parameter.

fix_gamma

TRUE if fix gamma (advection term), FALSE if estimate gamma.

theta_gamma_x

The x component of the advection term: `gamma_x = B_gamma_x

theta_gamma_yThe y component of the advection term: `gamma_y = B_gamma_y B_gamma_xThe design matrix for the x component of the advection term.B_gamma_yThe design matrix for the y component of the advection term.B_gamma_x_listA list of design matrices for the x component of the advection term on every time node, length(B_gamma_x_list) == nt-1.B_gamma_y_listA list of design matrices for the y component of the advection term on every time node, length(B_gamma_y_list) == nt-1.kappaKappa parameter from Matern SPDE.stabilizationTRUE if using a stabilization term (for implicit Euler)....Additional arguments (ignored).cParameter. ngme_operator object. Given a spatial and temporal model, build a non-separable space-time model.