Test ngme function

Test ngme function for different models

test_ngme(
  model,
  n_obs_per_rep,
  n_replicate = 1,
  control_opt = control_opt(),
  numer_grad = TRUE,
  f_noise = noise_nig(mu = -2, sigma = 1.5, nu = 0.5),
  n_gibbs_samples = 5,
  family = "normal",
  max.n = 1000,
  debug = FALSE,
  debug_f = FALSE,
  start = NULL,
  fix_theta_K = FALSE,
  fix_theta_mu = FALSE,
  fix_theta_sigma = FALSE,
  fix_nu = FALSE,
  seed = Sys.time()
)

Arguments

model

model name

n_obs_per_rep

number of observation per replicate

n_replicate

number of replicate

numer_grad

numerical gradient

max.n

maximum number for building mesh

debug

debug mode

debug_f

debug mode for latent process

n_parallel_chain

number of parallel chains

stop_points

number of stop points for convergence check

preconditioner

preconditioner, can be c("none", "fast", "full") "none" means no preconditioner, "fast" means precondition everything except for the parameter of K matrix (for speed reason), "full" means precondition everything

sampling_strategy

subsampling method of replicates of model, c("all", "is")

precond_by_diff_chain

logical, if TRUE, use different chains to estimate preconditioner (only computed at check points), if FALSE, use the same chain to estimate preconditioner (computed at each iteration)

compute_precond_each_iter

logical, if TRUE, compute preconditioner at each iteration, if FALSE, only compute preconditioner at check points (if has only 1 chain running, it will be set TRUE)