Generate control specifications for ngme() function.

These are configurations for ngme optimization process.

control_opt(
  seed = Sys.time(),
  burnin = 100,
  iterations = 500,
  estimation = TRUE,
  standardize_fixed = TRUE,
  stop_points = 10,
  iters_per_check = iterations/stop_points,
  n_parallel_chain = 4,
  max_num_threads = n_parallel_chain,
  exchange_VW = TRUE,
  n_slope_check = 3,
  std_lim = 0.1,
  trend_lim = 0.01,
  print_check_info = FALSE,
  preconditioner = "none",
  precond_eps = 1e-05,
  precond_by_diff_chain = FALSE,
  compute_precond_each_iter = FALSE,
  optimizer = adamW(),
  max_relative_step = 0.5,
  max_absolute_step = 0.5,
  reduce_var = FALSE,
  reduce_power = 0.75,
  threshold = 1e-05,
  window_size = 1,
  rao_blackwellization = FALSE,
  n_trace_iter = 10,
  sampling_strategy = "all",
  verbose = FALSE
)

Arguments

seed

set the seed for pesudo random number generator

burnin

interations for burn-in periods (before optimization)

iterations

optimization iterations

estimation

run the estimation process (call C++ in backend)

standardize_fixed

whether or not standardize the fixed effect

stop_points

number of stop points for convergence check (or specify iters_per_check)

iters_per_check

run how many iterations between each check point (or specify stop_points)

n_parallel_chain

number of parallel chains

max_num_threads

maximum number of threads used for parallel computing, by default will be set same as n_parallel_chain. If it is more than n_parallel_chain, the rest will be used to parallel different replicates of the model.

exchange_VW

exchange last V and W in each chian

n_slope_check

number of stop points for regression

std_lim

maximum allowed standard deviation

trend_lim

maximum allowed slope

print_check_info

print the convergence information

preconditioner

preconditioner, can be c("none", "fast", "full") "none" means no preconditioner, "fast" means precondition everything except for the parameter of K matrix (for speed reason), "full" means precondition everything

precond_eps

numerical, the gap used for estimate preconditioner, default is 1e-5

precond_by_diff_chain

logical, if TRUE, use different chains to estimate preconditioner (only computed at check points), if FALSE, use the same chain to estimate preconditioner (computed at each iteration)

compute_precond_each_iter

logical, if TRUE, compute preconditioner at each iteration, if FALSE, only compute preconditioner at check points (if has only 1 chain running, it will be set TRUE)

optimizer

choose different sgd optimization method, currently support "vanilla", "momentum", "adagrad", "rmsprop", "adam", "adamW" see ?vanilla, ?momentum, ?adagrad, ?rmsprop, ?adam, ?adamW

max_relative_step

max relative step allowed in 1 iteration

max_absolute_step

max absolute step allowed in 1 iteration

reduce_var

logical, reduce variace

reduce_power

numerical the power of reduce level

threshold

till when start to reduce the variance

window_size

numerical, length of window for final estimates

rao_blackwellization

use rao_blackwellization

n_trace_iter

use how many iterations to approximate the trace (Hutchinson’s trick)

sampling_strategy

subsampling method of replicates of model, c("all", "is") "all" means using all replicates in each iteration, "ws" means weighted sampling (each iteration use 1 replicate to compute the gradient, the sample probability is proption to its number of observations)

verbose

print estimation

Value

list of control variables

Details

To enable convergence check, we need multiple chains running. We compare the trend of the estimated parameter of length n_slope_check (linear regression) with trend_lim. We compare the standard devation of estimated parameters (in different chains) with std_lim.