Specifying a latent process model (wrapper function for each model)

Function used for defining of smooth and spatial terms within ngme model formulae. The function is a wrapper function for specific submodels. (see ngme_models_types() for available models).

Usage

f(
  index = NULL,
  model = "ar1",
  replicates = NULL,
  noise = noise_normal(),
  control = ngme_control_f(),
  data = NULL,
  A = NULL,
  A_pred = NULL,
  theta_K = NULL,
  W = NULL,
  fix_W = NULL,
  fix_theta_K = NULL,
  index_pred = NULL,
  debug = NULL,
  index_NA = NULL,
  ...
)

Arguments

index

symbol or numerical value: index or covariates to build index

model

1. string: type of model, 2. ngme.spde object

replicates

Representing the replicates

noise

1. string: type of model, 2. ngme.noise object (can also be specified in each ngme model)

control

control variables for f model

data

specifed or inherit from ngme formula

A

A Matrix connecting observation and mesh

A_pred

A Matrix connecting NA location and mesh

theta_K

Unbounded parameter for K

W

starting value of the process

fix_W

stop sampling for W

fix_theta_K

fix the estimation for theta_K.

index_pred

index for prediction

debug

Debug mode

index_NA

Logical vector, same as is.na(response var.)

...

additional arguments inherit the data from ngme function

Value

a list latent_in for constructing latent model, e.g. A, h, C, G, which also has 1. Information about K matrix 2. Information about noise 3. Control variables