ngme noise specification

Function for specifying ngme noise. Please use noise_nig and noise_normal for simpler usage. Use ngme_noise_types() to check all the available types.

ngme_noise(
  noise_type,
  mu = 0,
  sigma = 1,
  nu = 1,
  theta_mu = NULL,
  B_mu = NULL,
  theta_sigma = NULL,
  B_sigma = NULL,
  theta_sigma_normal = NULL,
  B_sigma_normal = NULL,
  fix_theta_mu = FALSE,
  fix_theta_sigma = FALSE,
  fix_nu = FALSE,
  V = NULL,
  fix_V = FALSE,
  single_V = FALSE,
  share_V = FALSE,
  corr_measurement = FALSE,
  index_corr = NULL,
  rho = 0,
  ...
)

noise_normal(
  sigma = NULL,
  theta_sigma = NULL,
  B_sigma = matrix(1),
  corr_measurement = FALSE,
  index_corr = NULL,
  ...
)

noise_nig(
  mu = NULL,
  sigma = NULL,
  nu = NULL,
  V = NULL,
  theta_mu = NULL,
  theta_sigma = NULL,
  B_mu = matrix(1),
  B_sigma = matrix(1),
  corr_measurement = FALSE,
  index_corr = NULL,
  ...
)

noise_gal(
  mu = NULL,
  sigma = NULL,
  nu = NULL,
  V = NULL,
  theta_mu = NULL,
  theta_sigma = NULL,
  B_mu = matrix(1),
  B_sigma = matrix(1),
  corr_measurement = FALSE,
  index_corr = NULL,
  ...
)

noise_normal_nig(
  sigma_normal = NULL,
  mu = NULL,
  sigma_nig = NULL,
  nu = NULL,
  V = NULL,
  theta_mu = NULL,
  theta_sigma_nig = NULL,
  theta_sigma_normal = NULL,
  B_mu = matrix(1),
  B_sigma_nig = matrix(1),
  B_sigma_normal = matrix(1),
  corr_measurement = FALSE,
  index_corr = NULL,
  ...
)

Arguments

noise_type

type of noise, "nig", "normal"

mu

specify the NIG noise parameter mu, see ?nig

sigma

specify the noise parameter sigma, see ?nig

nu

specify the NIG noise parameter nu (nu>0), see ?nig

theta_mu

specify a non-stationary noise using theta_mu

B_mu

Basis matrix for mu (if non-stationary)

theta_sigma

specify a non-stationary noise using theta_sigma

B_sigma

Basis matrix for sigma (if non-stationary)

theta_sigma_normal

for normal nosie with nig noise sharing same parameter

B_sigma_normal

for normal nosie with nig noise sharing same parameter

fix_theta_mu

fix the parameter of theta_mu

fix_theta_sigma

fix the parameter of theta_sigma

fix_nu

fix the parameter of nu

V

start value for V

fix_V

fix the sampling of V

single_V

TRUE if V is a single number

share_V

used only for bivariate model

corr_measurement

TRUE if we use correlated measurement noise

index_corr

used when corr_measurement=TRUE, indicate which observation has correlation

rho

used when corr_measurement=TRUE, starting point for correlation

...

additional arguments

sigma_normal

for normal nosie with nig noise sharing same parameter

sigma_nig

similar to sigma_normal

theta_sigma_nig

similar to theta_sigma_normal

B_sigma_nig

similar to B_sigma_nig

Value

a list of specification of noise

Details

The parameterization is given in ?nig. Moreover, for specifying non-stationary mu and sigma, $$\mu = B_{\mu} \theta_{\mu},$$ and $$\sigma = \exp (B_{\sigma} \theta_{\sigma}),$$

Examples

noise_normal(n = 10, sigma = 2)
#> Noise type: NORMAL
#> Noise parameters: 
#>     sigma = 2
noise_nig(mu = 1, sigma = 2, nu = 1, n=10)
#> Noise type: NIG
#> Noise parameters: 
#>     mu = 1
#>     sigma = 2
#>     nu = 1
noise_gal(mu = 1, sigma = 2, nu = 1, n=10)
#> Noise type: GAL
#> Noise parameters: 
#>     mu = 1
#>     sigma = 2
#>     nu = 1