Ngme space-time non-separable model specification

Usage

spacetime(
  mesh,
  lambda = 1,
  alpha = 2,
  cc = 1,
  kappa = 1,
  fix_gamma = FALSE,
  theta_gamma_x = 0,
  theta_gamma_y = 0,
  shared_theta_gamma = FALSE,
  B_gamma_x = matrix(1, nrow = mesh[[2]]$n, ncol = 1),
  B_gamma_y = matrix(1, nrow = mesh[[2]]$n, ncol = 1),
  B_gamma_x_list = NULL,
  B_gamma_y_list = NULL,
  stabilization = TRUE
)

Arguments

mesh

A list of two objects:

mesh_t - The temporal mesh

mesh_s - The spatial mesh

lambda

The spatial damping parameter.

alpha

2 or 4, SPDE smoothness parameter.

cc

Parameter c in the SPDE.

kappa

Kappa parameter from Matern SPDE.

fix_gamma

TRUE if fix gamma (advection term), FALSE if estimate gamma.

theta_gamma_x

The x component of the advection term: `gamma_x = B_gamma_x

theta_gamma_yThe y component of the advection term: `gamma_y = B_gamma_y shared_theta_gammaTRUE if share the same theta_gamma for all time nodes. (theta_gamma_x and theta_gamma_y will be the same)B_gamma_xThe design matrix for the x component of the advection term.B_gamma_yThe design matrix for the y component of the advection term.B_gamma_x_listA list of design matrices for the x component of the advection term on every time node, length(B_gamma_x_list) == nt-1.B_gamma_y_listA list of design matrices for the y component of the advection term on every time node, length(B_gamma_y_list) == nt-1.stabilizationTRUE if using a stabilization term (for implicit Euler). ngme_operator object. Implements a non-separable space-time model based on the advection-diffusion SPDE with first-order derivative in time. The model combines temporal and spatial components through a finite difference method (implicit Euler) for temporal discretization and finite element method (continuous Galerkin) for spatial discretization. When advection dominates diffusion, the "Streamline Diffusion" stabilization technique is applied. The model is particularly useful for large space-time datasets in environmental science, offering computationally efficient methods for parameter estimation, kriging prediction, and conditional simulations.For details, see https://www.sciencedirect.com/science/article/pii/S2211675324000381